add rake task to import existing tokens from CSV file
Now that we are ready to go live on this app, we need a way to
import the existing, previously issued tokens from the current
csv file.

This commit adds a rake task (rake import:tokens FILE=/path/to/file)
to do that. There were some issues around validations, as some of the
data in the existing Google doc is incomplete, yet we still need
those tokens imported. To make this work, I have:

- added a created_from field to the Token model ('web' or 'import')
- made some of the Token validations conditional on that value.