Consolidate ApiEnv and ProvisioningKey as 1 model
ApiEnv was an array, dressed up as a model, and it was not possible to create new envs within the application.
The ability to create new environments will soon be required.
We considered convertin ApiEnv to a full model.
This would have meant a 1:1 relationship between ApiEnv and ProvisioningKey.
The only attribute on ApiEnv would have been #name
We felt, therefore, that it would make more sense to consolidate ApiEnv and ProvisioningKey into one model.

- Renamed ProvisioningKey to Environment
- Token and AccessRequest now belong to Environment
- Environment has many Tokens and AccessRequests
- Fixed broken tests and tested migration
- New cuke is commented out due to the way in which Cucumber is being used at present, which means cucumber scenarios write to the dev DB and database cleaner does not clean.